# catalog/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Вывод всех категорий",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Массив категорий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Вывод категории с подкатегориями",
                "operationId": "ff3ab984d3fe877eb16da6009394d694",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Массив категории с подкатегорями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/admin/categories/create": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Создание новой категории",
                "operationId": "c8e6d9d4e3a02d178cf3d8c461785cb6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Игрушки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Массив с данными категории",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/admin/categories/{category}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Обновление категории",
                "operationId": "decc152d00e60f86768977bbbe517a13",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Игрушки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Массив с данными категории",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Удаление категории",
                "operationId": "b12ec127627519cbc7b97b203a92ce55",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория удалена"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "description": "Категория для продуктов",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID категории",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string",
                        "example": "Футболка"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "description": "Класс продукта",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "sub_category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID продукта",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "text",
                        "example": "Хлопок, сертифицированный по экологическим стандартам, при его производстве не используются агрессивные химикаты. Выбирая такой товар, вы проявляете заботу о себе и о планете."
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "integer",
                        "example": "9 990"
                    },
                    "previousPrice": {
                        "description": "Прежняя цена продукта",
                        "type": "integer",
                        "example": "10 990"
                    },
                    "sub_category_id": {
                        "description": "ID подктаегории товара",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
            "SubCategory": {
                "description": "Подкатегория для продуктов",
                "required": [
                    "id",
                    "name",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID подкатегории",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название подкатегории",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "category_id": {
                        "description": "Основная категория для подкатегории",
                        "type": "integer",
                        "example": "3"
                    },
                    "products": {
                        "description": "ID подкатегории товара",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        },
                        "example": [
                            {
                                "id": 1,
                                "name": "Худи",
                                "description": "Ea ut consequatur natus sed occaecati hic. Adipisci repellendus id cum itaque quis aliquid. Aut sunt suscipit in sequi tenetur. Sit illo quo velit quia quia possimus.",
                                "price": "KZT 34,227.00",
                                "previousPrice": "KZT 92,848.00",
                                "sub_category": "Верхняя одежда",
                                "category_id": "Одежда"
                            },
                            {
                                "id": 2,
                                "name": "Худи",
                                "description": "Suscipit et modi magnam in dolorum. Sit deserunt sunt aut non. Nulla nihil veritatis aut fuga consectetur quidem quae. Numquam earum excepturi sapiente et illo nisi sint.",
                                "price": "KZT 64,807.00",
                                "previousPrice": "KZT 42,079.00",
                                "sub_category": "Верхняя одежда",
                                "category_id": "Одежда"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "description": "Категория для продуктов",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID категории",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string",
                        "example": "Одежда"
                    },
                    "sub_categories": {
                        "description": "Подкатегорий основной категории",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubCategory"
                        },
                        "example": [
                            {
                                "id": "1",
                                "name": "Верхняя одежда"
                            },
                            {
                                "id": "2",
                                "name": "Штаны"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}